feat:개인 러닝보상 컴포넌트 분리 및 크루 주간 보상으로 재활용
