+++ web/app/crew-reward/page.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import RunRewardPage from '@/components/running/RunRewardPage';
+
+function Page() {
+  return <RunRewardPage type={'crew'} />;
+}
+
+export default Page;
